+++ app/src/main/AndroidManifest.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="io.jeffchang.splitdiff">
+    package="io.jeffchang.challenge">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="io.jeffchang.splitdiff.ChallengeApplication"
+        android:name="io.jeffchang.challenge.ChallengeApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -15,7 +15,7 @@
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
         <activity
-            android:name="io.jeffchang.splitdiff.MainActivity"
+            android:name="io.jeffchang.challenge.MainActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
